These are the source code for the utilities
This utility parses an XML file and produces "tree" output of all the values found. Grep the result for the variable you need, put it in to a file and include it in to the script to set a whole lot of variables.
You can use the split utility to precisely control which block of values you wish to import (i.e. find a record and then skip to the next start of record, capturing all its attributes).